requirements.txt
google-genai>=1.0.0
fpdf2>=2.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
python-dotenv>=1.0.0
